edition nom et coef ue

+++ Model/Sources/Model/Markable.swift
@@ -9,7 +9,7 @@ import Foundation
 
 public protocol Markable{
     
-    var Coef : Float {get}
+    var Coef : Int {get}
     var Mark : Float {get}
     
 }
